refactor: cli::cli is now cli::subcommands
