+++ src/cli/mod.rs
@@ -1,5 +1,5 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
-pub mod cli;
 pub mod humantoken;
 pub mod main;
+pub mod subcommands;
