+++ src/cli/subcommands/auth_cmd.rs
