+++ src/cli/subcommands/fetch_params_cmd.rs
@@ -5,7 +5,7 @@ use crate::shim::{sector::SectorSize, Inner};
 use crate::utils::proofs_api::paramfetch::{get_params_default, SectorSizeOpt};
 
 use super::cli_error_and_die;
-use crate::cli::cli::Config;
+use crate::cli::subcommands::Config;
 
 #[allow(missing_docs)]
 #[derive(Debug, clap::Args)]
