+++ src/cli/subcommands/sync_cmd.rs
@@ -14,7 +14,7 @@ use clap::Subcommand;
 use ticker::Ticker;
 
 use super::Config;
-use crate::cli::cli::{format_vec_pretty, handle_rpc_err};
+use crate::cli::subcommands::{format_vec_pretty, handle_rpc_err};
 
 #[derive(Debug, Subcommand)]
 pub enum SyncCommands {
